requirements.txt
kernels==0.3.0
numpy==1.21.6
opencv_python==4.8.0.74
Pillow==9.5.0
scikit_learn==1.0.2
scipy==1.7.3
timm==0.4.12
torch==1.13.0
torchvision==0.14.0
